pyproject.toml
[project]
name = "brain-gnn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytorch-lightning[extra]>=2.4.0",
    "torch-geometric>=2.6.1",
    "torchvision>=0.20.1",
    "wandb>=0.19.1",
]

[dependency-groups]
dev = [
    "ruff>=0.8.3",
]
